+++ src/app/dashboard/components/sidebar-wide.tsx
