+++ src/app/dashboard/profile/avatar-name.tsx
@@ -1,6 +1,4 @@
 import Image from 'next/image';
-import { createClient } from '@/lib/supabase/server';
-import { Icon } from '@radix-ui/react-select';
 
 export default function AboutStoreTile() {
   return (
